Исправления

+++ spring-05/src/main/resources/schema.sql
@@ -7,8 +7,8 @@ create table author(id IDENTITY primary key, first_name varchar(255) not null, l
 drop table if exists book;
 CREATE TABLE book(id IDENTITY primary key,
     name varchar(255) not null,
-    author_id INT,
-    genre_id INT,
+    author_id BIGINT,
+    genre_id BIGINT,
     foreign key (author_id) references author(id),
     foreign key (genre_id) references genre(id)
 )
